\name{CellProfileR-internal}
\alias{center.on.median}
\alias{checkValidFilenames}
\alias{readPlateMapping}
\alias{G144_plate_map_file}
\alias{example_plate_map_file}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{CellProfileR-internal}
\description{Internal functions}
\usage{}

\arguments{}
\details{
These are all internal package material:
\itemize{
    \item center.on.median
    \item checkValidFilenames
    \item readPlateMapping
    \item example_plate_map_file -- example 96 well mapping file
}

}
\value{}
\references{}
\author{Amos Folarin}
\note{}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{}
\examples{}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
