% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df.R
\name{clean_df}
\alias{clean_df}
\title{Clean raw data}
\usage{
clean_df(df)
}
\arguments{
\item{df}{Dataframe of raw Cell Profiler data.}
}
\value{
A dataframe with only the necessary plate, well, count, median and
    image quality columns.
}
\description{
Remove redundant columns from the raw Cell Profiler data.
}
