% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_mp.R
\name{clean_mp}
\alias{clean_mp}
\title{Clean MP data}
\usage{
clean_mp(mp_df, sep, tx_col = "tx")
}
\arguments{
\item{mp_df}{Dataframe output from \code{calc_mp}, containing mp-values and
treatment and batch details.}

\item{sep}{Symbol separating the treatment and concentration in the 
concatenated column, as string.}

\item{tx_col}{Name of the column containing the concatenated treatment and
concentration data.}
}
\description{
If each unique treatment and concentration was treated as a separate 
treatment in \code{calc_mp}, the "tx" column of the output would contain the 
treatment and concentration concatenated together. This function takes the 
dataframe output by \code{calc_mp} and outputs the dataframe with two extra 
columns; 'conc' which contains just the concentration and 'compound', which 
contains just the compound name.
}
\details{
Note that this function assumes that each concatenated string has the 
treatment first and then the concentration. It also assumes that the 
concentration is numerical (does not contain any non-numerical characters) 
and does not contain the 'sep' symbol.
}
