% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CP_data}
\alias{CP_data}
\title{Raw CellProfiler data}
\format{An object of class \code{data.frame} with 1600 rows and 371 columns.}
\usage{
CP_data
}
\description{
Raw CellProfiler data from imaging cells in a 384 (24 by 16) well plate. 
Cells in each well were exposed to a different compound, left to grow and
the imaged using three stains; CMFDA, DAPI and Phalloidin. CellProfiler was 
then used to extract phenotype data from each image.
}
\keyword{datasets}
