% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_features.R
\name{filter_NA}
\alias{filter_NA}
\alias{filter_lowVar}
\alias{filter_cor}
\title{Filter features}
\usage{
filter_NA(df, NA_cutoff, out_cols = FALSE)

filter_lowVar(df, freqCut = 95/5, uniqueCut = 10, out_cols = FALSE)

filter_cor(df, cor_cutoff, out_cols = FALSE)
}
\arguments{
\item{df}{Dataframe of Cell Profiler data to be filtered.}

\item{NA_cutoff}{Columns with a total number of NAs or NaNs greater than
the NA_cutoff number will be removed.}

\item{out_cols}{Single logical indicating whether the filtered column names
should be output as well.}

\item{freqCut}{Passed to \code{caret::nearZeroVar}. The cutoff for the ratio 
of the most common value to the second most common value.}

\item{uniqueCut}{Passed to \code{caret::nearZeroVar}. The cutoff for the 
percentage of distinct values out of the number of total samples.}
}
\value{
A dataframe with the filtered columns removed. If out_cols is set
    to TRUE, a list of 2, where the 1st element is the filtered dataframe
    and the 2nd element is a vector of the filtered column names.
}
\description{
Filters columns (features) of Cell Profiler data to exclude any columns 
with NA/NaNs, low variance or are highly correlated.
}
\section{Functions}{
\itemize{
\item \code{filter_NA}: Filters columns with any NA/NaN's or a total number 
above a certain number if 'NA_cutoff' given.

\item \code{filter_lowVar}: Filters columns with low variance (e.g. if entire 
column consists of the same value) using the \code{caret::nearZeroVar} 
function. See 
\href{https://cran.r-project.org/web/packages/caret/caret.pdf}{caret 
documentation} for details on this function.

\item \code{filter_cor}: Calculates the Pearson correlation matrix for the
'Median' columns then filters columns that are highly correlated to each 
other using \code{caret::findCorrelation}. Note that you should run 
\code{filter_lowVar} on your data before filtering for high correlations.
}}

