% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_images.R
\name{filterImages}
\alias{filterImages}
\alias{plotFiltered}
\title{Filter images}
\usage{
filterImages(df, num_IQR)

plotFiltered(filtered_df, num_images, wells = 384)
}
\arguments{
\item{df}{Dataframe of raw Cell Profiler data, where each row is 1 image
NOT one well.}

\item{num_IQR}{Number of IQR's below the 25% quantile to place cutoff 
threshold.}

\item{filtered_df}{Dataframe of Cell Profiler data that has been filtered
for poor quality images (rows).}

\item{num_images}{The number of images taken per well.}

\item{wells}{The number of wells per plate}
}
\value{
\itemize{
    \item \code{filterImages} returns the dataframe with poor quality images
    (rows) removed.
    \item \code{plotFiltered} creates a bar plot in the current graphics
    device.
}
}
\description{
Filter poor quality images (using the powerloglog slope feature). 
\code{plotFiltered} plots the number of images filtered for each plate.
}
\section{Functions}{
\itemize{
\item \code{filterImages}: Filter images (rows) with a powerloglog slope 
value below the threshold ('num_IQR' * IQR below the 25th percentile) in 
ANY channel.

\item \code{plotFiltered}: Creates bar graph of the number of images filtered 
for each plate.
}}

