\name{GBMDrugScreen}
\docType{data}
\alias{GBMDrugScreen}
\alias{per.img}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
CellProfiler data from a glioblastoma celline drug screen (see vignette
data("workflow3")).
}
\description{
A part of a compound screen with the objective of identifying compounds that
are cytotoxic or cytostatic to the tumour cells.

The screen was carried out on 6 x 96 well plate. 1) Phase contrast images were
taken to assess total cell numbers and metaphase stage cells. This was then
analysed on a CellProfiler pipeline to count the number of images on each image
type.
}
\usage{
data("GBMDrugScreen")
}
\format{
GBMDrugScreen is a dataset that holds a the data.frame per.img.s represenative
of the per_image table generated by CellProfiler.

The data.frame per.img.s has 576 rows 122 columns.   

\enumerate{
\item "ImageNumber"                                                     
\item "Image_Count_CellAndNuclei"                                       
\item "Image_FileName_Phase"                                            
\item "Image_Group_Index"                                               
\item "Image_Group_Number"                                              
\item "Image_MD5Digest_Phase"                                           
\item "Image_Metadata_Hour"                                             
\item "Image_Metadata_ImagesPerWell"                                    
\item "Image_Metadata_Plate"                                            
\item "Image_Metadata_Well"                                             
\item "Image_PathName_Phase"                                            
\item "Image_Texture_AngularSecondMoment_ImageAfterExpo2_3"             
\item "Image_Texture_Contrast_ImageAfterExpo2_3"                        
\item "Image_Texture_Correlation_ImageAfterExpo2_3"                     
\item "Image_Texture_DifferenceEntropy_ImageAfterExpo2_3"               
\item "Image_Texture_DifferenceVariance_ImageAfterExpo2_3"              
\item "Image_Texture_Entropy_ImageAfterExpo2_3"                         
\item "Image_Texture_Gabor_ImageAfterExpo2_3"                           
\item "Image_Texture_InfoMeas1_ImageAfterExpo2_3"                       
\item "Image_Texture_InfoMeas2_ImageAfterExpo2_3"                       
\item "Image_Texture_InverseDifferenceMoment_ImageAfterExpo2_3"         
\item "Image_Texture_SumAverage_ImageAfterExpo2_3"                      
\item "Image_Texture_SumEntropy_ImageAfterExpo2_3"                      
\item "Image_Texture_SumVariance_ImageAfterExpo2_3"                     
\item "Image_Texture_Variance_ImageAfterExpo2_3"                        
\item "Image_Threshold_FinalThreshold_CellAndNuclei"                    
\item "Image_Threshold_FinalThreshold_ThreshMathInvert"                 
\item "Image_Threshold_OrigThreshold_CellAndNuclei"                     
\item "Image_Threshold_OrigThreshold_ThreshMathInvert"                  
\item "Image_Threshold_SumOfEntropies_CellAndNuclei"                    
\item "Image_Threshold_SumOfEntropies_ThreshMathInvert"                 
\item "Image_Threshold_WeightedVariance_CellAndNuclei"                  
\item "Image_Threshold_WeightedVariance_ThreshMathInvert"               
\item "Mean_CellAndNuclei_Number_Object_Number"                         
\item "Mean_CellAndNuclei_AreaShape_Area"                               
\item "Mean_CellAndNuclei_AreaShape_Eccentricity"                       
\item "Mean_CellAndNuclei_AreaShape_EulerNumber"                        
\item "Mean_CellAndNuclei_AreaShape_Extent"                             
\item "Mean_CellAndNuclei_AreaShape_FormFactor"                         
\item "Mean_CellAndNuclei_AreaShape_MajorAxisLength"                    
\item "Mean_CellAndNuclei_AreaShape_MinorAxisLength"                    
\item "Mean_CellAndNuclei_AreaShape_Orientation"                        
\item "Mean_CellAndNuclei_AreaShape_Perimeter"                          
\item "Mean_CellAndNuclei_AreaShape_Solidity"                           
\item "Mean_CellAndNuclei_AreaShape_Zernike_0_0"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_1_1"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_2_0"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_2_2"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_3_1"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_3_3"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_4_0"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_4_2"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_4_4"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_5_1"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_5_3"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_5_5"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_6_0"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_6_2"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_6_4"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_6_6"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_7_1"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_7_3"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_7_5"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_7_7"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_8_0"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_8_2"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_8_4"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_8_6"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_8_8"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_9_1"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_9_3"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_9_5"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_9_7"                        
\item "Mean_CellAndNuclei_AreaShape_Zernike_9_9"                        
\item "Mean_CellAndNuclei_Intensity_IntegratedIntensityEdge_ImgAftrExp2"
\item "Mean_CellAndNuclei_Intensity_IntegratedIntensity_ImageAfterExpo2"
\item "Mean_CellAndNuclei_Intensity_LowerQuartileIntensity_ImagAftrExp2"
\item "Mean_CellAndNuclei_Intensity_MassDisplacement_ImageAfterExpo2"   
\item "Mean_CellAndNuclei_Intensity_MaxIntensityEdge_ImageAfterExpo2"   
\item "Mean_CellAndNuclei_Intensity_MaxIntensity_ImageAfterExpo2"       
\item "Mean_CellAndNuclei_Intensity_MeanIntensityEdge_ImageAfterExpo2"  
\item "Mean_CellAndNuclei_Intensity_MeanIntensity_ImageAfterExpo2"      
\item "Mean_CellAndNuclei_Intensity_MedianIntensity_ImageAfterExpo2"    
\item "Mean_CellAndNuclei_Intensity_MinIntensityEdge_ImageAfterExpo2"   
\item "Mean_CellAndNuclei_Intensity_MinIntensity_ImageAfterExpo2"       
\item "Mean_CellAndNuclei_Intensity_StdIntensityEdge_ImageAfterExpo2"   
\item "Mean_CellAndNuclei_Intensity_StdIntensity_ImageAfterExpo2"       
\item "Mean_CellAndNuclei_Intensity_UpperQuartileIntensity_ImagAftrExp2"
\item "Mean_CellAndNuclei_Location_Center_X"                            
\item "Mean_CellAndNuclei_Location_Center_Y"                            
\item "Mean_CellAndNuclei_RadialDistribution_FracAtD_ImageAfterExp2_1f6"
\item "Mean_CellAndNuclei_RadialDistribution_FracAtD_ImageAfterExp2_2f6"
\item "Mean_CellAndNuclei_RadialDistribution_FracAtD_ImageAfterExp2_3f6"
\item "Mean_CellAndNuclei_RadialDistribution_FracAtD_ImageAfterExp2_4f6"
\item "Mean_CellAndNuclei_RadialDistribution_FracAtD_ImageAfterExp2_5f6"
\item "Mean_CellAndNuclei_RadialDistribution_FracAtD_ImageAfterExp2_6f6"
\item "Mean_CellAndNuclei_RadialDistribution_MeanFrac_ImageAftrExp2_1f6"
\item "Mean_CellAndNuclei_RadialDistribution_MeanFrac_ImageAftrExp2_2f6"
\item "Mean_CellAndNuclei_RadialDistribution_MeanFrac_ImageAftrExp2_3f6"
\item "Mean_CellAndNuclei_RadialDistribution_MeanFrac_ImageAftrExp2_4f6"
\item "Mean_CellAndNuclei_RadialDistribution_MeanFrac_ImageAftrExp2_5f6"
\item "Mean_CellAndNuclei_RadialDistribution_MeanFrac_ImageAftrExp2_6f6"
\item "Mean_CellAndNuclei_RadialDistribution_RadialCV_ImageAftrExp2_1f6"
\item "Mean_CellAndNuclei_RadialDistribution_RadialCV_ImageAftrExp2_2f6"
\item "Mean_CellAndNuclei_RadialDistribution_RadialCV_ImageAftrExp2_3f6"
\item "Mean_CellAndNuclei_RadialDistribution_RadialCV_ImageAftrExp2_4f6"
\item "Mean_CellAndNuclei_RadialDistribution_RadialCV_ImageAftrExp2_5f6"
\item "Mean_CellAndNuclei_RadialDistribution_RadialCV_ImageAftrExp2_6f6"
\item "Mean_CellAndNuclei_Texture_AngularSecondMoment_ImageAfterExpo2_3"
\item "Mean_CellAndNuclei_Texture_Contrast_ImageAfterExpo2_3"           
\item "Mean_CellAndNuclei_Texture_Correlation_ImageAfterExpo2_3"        
\item "Mean_CellAndNuclei_Texture_DifferenceEntropy_ImageAfterExpo2_3"  
\item "Mean_CellAndNuclei_Texture_DifferenceVariance_ImageAfterExpo2_3" 
\item "Mean_CellAndNuclei_Texture_Entropy_ImageAfterExpo2_3"            
\item "Mean_CellAndNuclei_Texture_Gabor_ImageAfterExpo2_3"              
\item "Mean_CellAndNuclei_Texture_InfoMeas1_ImageAfterExpo2_3"          
\item "Mean_CellAndNuclei_Texture_InfoMeas2_ImageAfterExpo2_3"          
\item "Mean_CellAndNuclei_Texture_InverseDifferenceMoment_ImgAftrExp2_3"
\item "Mean_CellAndNuclei_Texture_SumAverage_ImageAfterExpo2_3"         
\item "Mean_CellAndNuclei_Texture_SumEntropy_ImageAfterExpo2_3"         
\item "Mean_CellAndNuclei_Texture_SumVariance_ImageAfterExpo2_3"        
\item "Mean_CellAndNuclei_Texture_Variance_ImageAfterExpo2_3"  

}

}

\source{
Unpublished data.
}

\references{
The GBMDrugScreen dataset was donated by Davide Danovi.
}

\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
 This dataset is used in the vignette for Workflow3.
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
