% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpvalue.R
\name{mpvalue}
\alias{mpvalue}
\title{mp-value}
\usage{
mpvalue(dataset, txlabels, batchlabels, datacols, negctrls,
  allbyall = FALSE, dirprefix, outfile = "all_mp-values.txt",
  loadingsout = FALSE, pcaout = FALSE, gammaout = FALSE)
}
\arguments{
\item{dataset}{Dataframe containing all data where each row is a sample and
each columns is a feature/variable.}

\item{txlabels}{Name of the column containing the treatment labels, as 
string.}

\item{batchlabels}{Name of the column containing the batch labels, as string}

\item{datacols}{Vector of column numbers containing the feature/variable data
to be used for calculating the mp-value.}

\item{negctrls}{The string identifying the negative control, in the 
'txlabels' column.}

\item{allbyall}{Logical indicating whether to do all treatment-treatment
comparisons (TRUE) or only treatment-control comparisons.}

\item{dirprefix}{Path to the directory to store results in, as string. 
This path will be created if it does not already exist.}

\item{outfile}{Name of the file to save all the mp-values, as string.}

\item{loadingsout}{Logical indicating whether the PCA loadings (eigenvectors)
should be output.}

\item{pcaout}{Logical indicating whether to output PCA values}

\item{gammaout}{Logical indicating whether to output gamma distribution
parameters, p-value of goodness of fit and sample p-value according to 
the fit distribution.
NOTE: When gammaout=TRUE, warnings like these may be displayed:}
}
\description{
Calculate mp-values as described in 
\href{https://journals.sagepub.com/doi/10.1177/1087057112469257}{Hutz et al.}.
}
