% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_all.R
\name{norm_all}
\alias{norm_all}
\title{Normalise data}
\usage{
norm_all(df, data_cols, count_col = "Count_Cells", compound_col, negCtrl,
  rescale_type = "robust")
}
\arguments{
\item{df}{Dataframe containing the CellProfiler data to normalise.}

\item{data_cols}{Numeric vector. Indicies of the data columns of df. All of 
these columns should be of numerice type. All remaining (non 'data_cols')
should NOT be of numeric type.}

\item{count_col}{Name of the column containing the raw cell counts, as 
string.}

\item{compound_col}{Name of the column in 'df' containing the compound 
details.}

\item{negCtrl}{Name of the negative control compound.}

\item{rescale_type}{Type of rescaling to use, as string. Will be passed to 
\code{reshape::rescaler}.}
}
\value{
A dataframe with the following columns, (in order): 
\itemize{
    \item Annotation columns (all non-data columns).
    \item Cell count columns: raw cell count (the \code{count_col} column 
    from \code{df}), cell count normalised to negative control, z score 
    normalised cell count (using \code{reshape::rescaler}). The latter two 
    columns are both calculated using the \code{count_col} column 
    from \code{df}. Note that there should only be ONE count column. If 
    \code{df} has more than one count column (e.g. 'count_cells' AND 
    'count_cells_unfiltered" columns), delete these such that only one 
    remains.
    \item All remaining data columns z score normalised (using 
    \code{reshape::rescaler}). All of these columns names will be the same 
    as the input dataframe.
}
}
\description{
Takes in CellProfiler morphology data and uses \code{reshape::rescaler} to 
normalise all data columns. Also calculates count value normalised to the
negative control, for each plate, and retains the raw count value.
}
