% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_differences.R
\name{plot_countFilt}
\alias{plot_countFilt}
\alias{df_countFilt}
\alias{calc_countsImageQC}
\alias{plot_countImageQC}
\alias{df_countImageQC}
\title{Count differences}
\usage{
plot_countFilt(df_raw)

df_countFilt(df_raw, n = 10)

calc_countsImageQC(df_raw, df_filtered)

plot_countImageQC(df_raw, df_filtered, CP_unfiltered = FALSE)

df_countImageQC(df_raw, df_filtered, n = 10, CP_unfiltered = FALSE)
}
\arguments{
\item{df_raw}{Dataframe of raw Cell Profiler data with no images (rows)
removed. Each row should be 1 image NOT 1 well.}

\item{n}{Number of rows (images), with the largest differences, to include
in the output dataframe.}

\item{df_filtered}{Dataframe of Cell Profiler data after filtering for poor
quality images (rows). Again, each row should be 1 image NOT 1 well.}

\item{CP_unfiltered}{Single logical indicating whether the differences 
(before and after image filtering) should be calculted and shown for the
CP filtered column 'Count_Cells' or CP unfiltered column 
'Count_Cells_unfiltered'.}
}
\description{
Determine the differences in count values between images (rows) before and
after Cell Profiler filtering or manual filtering of poor quality and 
present as dataframe or plot.
}
\section{Functions}{
\itemize{
\item \code{plot_countFilt}: Calculates the difference between 
'Count_Cells_unfiltered' and 'Count_Cells' (Cell Profiler filtering)
for each row (image) and creates a density plot of the differences for 
ALL plates.

\item \code{df_countFilt}: Calculates the difference between 
'Count_Cells_unfiltered' and 'Count_Cells' (Cell Profiler filtering) 
for each row (image) and outputs a dataframe containing the 'n' rows 
(images) that have the largest differences. Plate, well and image number 
also shown in dataframe.

\item \code{calc_countsImageQC}: Calculates the per well total count before and 
after filtering for poor quality images, for BOTH the 
'Count_Cells_unfiltered' and 'Count_Cells' column (CP unfiltered and
filtered). Mostly for internal use but made available as well.

\item \code{plot_countImageQC}: Calculates the difference in total count per well
before and after filtering for poor quality images (for either the 
'Count_Cells_unfiltered' or 'Count_Cells' column) then plots density 
plot of the differences for EACH plate.

\item \code{df_countImageQC}: Calculates the difference in total count per well
before and after filtering for poor quality images (for either the 
'Count_Cells_unfiltered' or 'Count_Cells' column) then produces a 
dataframe containing the 'n' rows (wells) with the largest differences.
}}

