% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plateHeat.R
\name{plot_plateHeat}
\alias{plot_plateHeat}
\title{Plot plate heatmaps}
\usage{
plot_plateHeat(df_PW, column, plate = 384)
}
\arguments{
\item{df_PW}{Dataframe of per well Cell Profiler data.}

\item{column}{Name of the column (feature) to plot, as string. Matching
for column name will be doing using grep but enough of the column name
must be given to identify a single column in the dataframe.}

\item{plate}{Number of wells in a plate.}
}
\description{
Plot a feature (column) as a plate heatmap for each plate. Scale is uniform
across all plates, such that the max value = red, min value = blue and 
median value = white.
}
