% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plls.R
\name{plot_plls}
\alias{plot_plls}
\title{Plot powerloglog slope}
\usage{
plot_plls(df, channel)
}
\arguments{
\item{df}{Dataframe containing a 'ImageQuality_PowerLogLogSlope' column}

\item{channel}{Name of the channel to be plotted, exactly as written in
Cell Profiler output, as string.}
}
\value{
One density plot for each plate in current graphics device.
}
\description{
Plot the density of the image quality metric, powerloglog slope for a 
specified channel, for each plate.
}
