% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAll.R
\name{readALL}
\alias{readALL}
\title{Read in & combine}
\usage{
readALL(paths, sep = ",")
}
\arguments{
\item{paths}{Character vector of folder paths the Cell Profiler data is in.
Function will use \code{list.files} on each folder path and read in
all files in each folder. Paths should NOT have a trailing "/".}

\item{sep}{The field separator character. Values on each line of the file 
are separated by this character.}
}
\description{
Read in all data output by Cell Profiler (e.g. for one cell line/experiment) 
and combine (bind rows) together into on data.table.
}
