\name{sqLiteConnect}
\alias{sqLiteConnect}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Create a connection to an SQLite CellProfiler database.
}
\description{
SQLite database connect - returns connection object to the SQLite database created
    by CellProfiler ExportToDatabase module.
}
\usage{
sqLiteConnect(database.file)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{database.file}{
    \code{database.file} the name of the SQLite database file generated by 
  CellProfiler (e.g. *.db) include the path where not in the working directory.
}
}
\details{
SQLite database connect - returns connection to the SQLite database created 
   by CellProfiler ExportToDatabase module.
}
\value{
returns a conection object to the sqlite database.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Amos Folarin
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%See Also as \code{DBI::dbConnect}, \code{DBI::dbDriver}.
}
\examples{
    
    #-----mysqlite example------<runnable>
     
    #get the filepath to the database file from the CellProfileR package
    db.file <- system.file("extdata","g144_db.sqlite", package="CellProfileR")
    
    #copy the sqlite database to a tempdir
    file.copy(db.file, tempdir(), overwrite = TRUE, recursive = TRUE)
    
    #connect to this sqlite database &  load into R the different tables stored 
    con <- sqLiteConnect(database.file=file.path(tempdir(),"g144_db.sqlite"))
    per.img <- loadTable(con, "per_image")
    
    
  
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

