\name{StemCellScreen}
\docType{data}
\alias{StemCellScreen}
\alias{per.img.s}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
CellProfiler data from a stem cell drug screen (see vignette data("workflow1")).
}
\description{
A part of a compound screen with the objective of identifying compounds that
modulated differentiation within of a stem cell line.

The screen was carried out on a 96 well plate. 2 images were acquired at each
timepoint, 1) Phase contrast images were taken to assess total cell numbers and
2) fluorescence images were taken to assess differentiation. This was then 
analysed on a CellProfiler pipeline to count the number of images on each image 
type.
}
\usage{
data("StemCellScreen")
}
\format{
StemCellScreen is a dataset that holds a the data.frame represenative of the
per_image table generated by CellProfiler.

The data.frame per.img has 2112 rows 48 columns.
\enumerate{
\item "ImageNumber"                                            
\item "Image_Count_GFPCells"                                   
\item "Image_Count_PhaseCells"                                 
\item "Image_FileName_GFPImage"                                
\item "Image_FileName_PhaseImage"                              
\item "Image_Group_Index"                                      
\item "Image_Group_Number"                                     
\item "Image_MD5Digest_GFPImage"                               
\item "Image_MD5Digest_PhaseImage"                             
\item "Image_Metadata_Hours"                                   
\item "Image_Metadata_Plate"                                   
\item "Image_Metadata_Well"                                    
\item "Image_PathName_GFPImage"                                
\item "Image_PathName_PhaseImage"                              
\item "Image_Threshold_FinalThreshold_GFPCells"                
\item "Image_Threshold_FinalThreshold_GFPThresh"               
\item "Image_Threshold_FinalThreshold_PhaseCells"              
\item "Image_Threshold_FinalThreshold_PhaseThresh"             
\item "Image_Threshold_OrigThreshold_GFPCells"                 
\item "Image_Threshold_OrigThreshold_GFPThresh"                
\item "Image_Threshold_OrigThreshold_PhaseCells"               
\item "Image_Threshold_OrigThreshold_PhaseThresh"              
\item "Image_Threshold_SumOfEntropies_GFPCells"                
\item "Image_Threshold_SumOfEntropies_GFPThresh"               
\item "Image_Threshold_SumOfEntropies_PhaseCells"              
\item "Image_Threshold_SumOfEntropies_PhaseThresh"             
\item "Image_Threshold_WeightedVariance_GFPCells"              
\item "Image_Threshold_WeightedVariance_GFPThresh"             
\item "Image_Threshold_WeightedVariance_PhaseCells"            
\item "Image_Threshold_WeightedVariance_PhaseThresh"           
\item "Mean_GFPCells_Intensity_IntegratedIntensityEdge_GFPgrey"
\item "Mean_GFPCells_Intensity_IntegratedIntensity_GFPgrey"    
\item "Mean_GFPCells_Intensity_LowerQuartileIntensity_GFPgrey" 
\item "Mean_GFPCells_Intensity_MassDisplacement_GFPgrey"       
\item "Mean_GFPCells_Intensity_MaxIntensityEdge_GFPgrey"       
\item "Mean_GFPCells_Intensity_MaxIntensity_GFPgrey"           
\item "Mean_GFPCells_Intensity_MeanIntensityEdge_GFPgrey"      
\item "Mean_GFPCells_Intensity_MeanIntensity_GFPgrey"          
\item "Mean_GFPCells_Intensity_MedianIntensity_GFPgrey"        
\item "Mean_GFPCells_Intensity_MinIntensityEdge_GFPgrey"       
\item "Mean_GFPCells_Intensity_MinIntensity_GFPgrey"           
\item "Mean_GFPCells_Intensity_StdIntensityEdge_GFPgrey"       
\item "Mean_GFPCells_Intensity_StdIntensity_GFPgrey"           
\item "Mean_GFPCells_Intensity_UpperQuartileIntensity_GFPgrey" 
\item "Mean_GFPCells_Location_Center_X"                        
\item "Mean_GFPCells_Location_Center_Y"                        
\item "Mean_PhaseCells_Location_Center_X"                      
\item "Mean_PhaseCells_Location_Center_Y"  
}

}

\source{
Unpublished data.
}

\references{
The StemCellScreen dataset was donated by Bart Baranowski.
}

\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
 This dataset is used in the vignette for Workflow1.
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
