% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_PerWell.R
\name{summarise_PerWell}
\alias{summarise_PerWell}
\title{Summarise Per well}
\usage{
summarise_PerWell(df, num_images, filtered = FALSE)
}
\arguments{
\item{df}{Dataframe of raw (per image) Cell Profiler data.}

\item{num_images}{The number of images taken per well.}

\item{filtered}{Single logical indicated whether poor quality images have 
been filtered (thus the number of images per well is variable).}
}
\value{
A dataframe or data.table (if filtered = TRUE) with one value per
    well.
}
\description{
Calculate the sum (for 'Count' features) or median (for 'Median' features)
value of the Cell Profiler columns (features) for each well.
}
