% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CP_data}
\alias{CP_data}
\title{Raw CellProfiler data}
\format{A data.table with 1600 rows (800 rows per plate), and 371 
variables:
\describe{
    \item{Metadata_Barcode}{Plate ID.}
    \item{Metadata_WellID}{Well ID. The plate is numbered such that the 
      columns go from 1-24 and the rows from A-P.}
    \item{Compound_ID}{The compound ID of the compound the well was exposed
      to. "DMSO" was the negative control compound.}
    \item{Concentration}{The concentration of the compound the well was
      exposed to. Several concentrations of the same compound was used for
      some compounds.}
    \item{CellProfiler output}{The remaining columns were output by 
      CellProfiler. Phenotype columns from CellProfiler have the general 
      structure: `MeasurementType_Location_Category_SpecificFeatureName` and
      are described in more detail below.}
}}
\usage{
CP_data
}
\description{
Raw CellProfiler data from imaging cells in a 384 (24 by 16) well plate. 
Cells in each well were exposed to a different compound, left to grow and
then imaged using three stains; CMFDA, DAPI and Phalloidin. CellProfiler was 
then used to extract phenotype data from images taken of each well.
}
\details{
Each row provides metatdata and information extracted by 
CellProfiler from one image taken of cells grown in one well. There were
four rows (images) taken per well. There were two 384 well plates. Only 200 
wells from each 384 well plate were included to minimise the size of the data 
set. A number of columns output by CellProfiler were also removed (including 
'Mean' and 'StDev' columns) to minimise the size of the data set.

Details of how the CellProfiler columns were named are as follows:
\describe{
    \item{MeasurementType}{This tells you the type of data contained in the
      measurement. Most columns are 'Median', which means that the median
      value of that measurement, for all cells identified in that image, is
      provided.}
    \item{Location}{The location of the cell this measurement is from. This
      is generally 'Nuclei', 'Cytoplasm' or 'Actin' (which captures  
      cell cytoskeleton details.)}
    \item{Category}{Description of the type of measurement taken. For 
      example 'AreaShape' provides information about the area or shape of
      a cell location and 'Texture' provides texture information about
      a cell location.}
    \item{SpecificFeatureName}{Details about the measurement made.}
}

More details about how each variable (column) is named and how each variable 
is calculated can be found in the 
\href{http://cellprofiler-manual.s3.amazonaws.com/CellProfiler-3.1.5/index.html}{CellProfiler documentation}.
}
\keyword{datasets}
