% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_images.R
\name{filterImages}
\alias{filterImages}
\alias{plotFiltered}
\title{Filter images}
\usage{
filterImages(df, num_IQR)

plotFiltered(filtered_df, num_images, annot, num_wells)
}
\arguments{
\item{df}{Dataframe of raw CellProfiler data, where each row is 1 image
NOT one well.}

\item{num_IQR}{Number of IQR's below the 25% quantile to place cutoff 
threshold.}

\item{filtered_df}{The plate ID column of the dataframe of CellProfiler 
data that has been filtered for poor quality data, as a dataframe or 
vector. E.g. if the plate ID column from the filtered dataframe 
\code{filtered_df} is called 'Metadata_Barcode', 
\code{filtered_df[,'Metadata_Barcode']} OR 
\code{filtered_df[['Metadata_Barcode']]} can be used.}

\item{num_images}{The number of images taken per well.}

\item{annot}{Optional argument. The plate ID column from the annotation data,
as a vector or dataframe. E.g. if the plate ID column from the 
\code{annot} dataframe is called 'VCFG_Compound_Plate_ID', 
\code{annot[,'VCFG_Compound_Plate_ID']} OR
\code{annot[['VCFG_Compound_Plate_ID']]} can be used. If given, the 
number of filtered images (rows) will be calculated as: 
(number of used wells) * \code{num_images} - (number of images remaining
after filtering). If not given, it will be assumed that all wells were 
used.}

\item{num_wells}{Number of wells per plate.}
}
\value{
\itemize{
    \item \code{filterImages} returns the dataframe with poor quality images
    (rows) removed.
    \item \code{plotFiltered} creates a bar plot in the current graphics
    device.
}
}
\description{
Filter poor quality images (using the powerloglog slope feature). 
\code{plotFiltered} plots the number of images filtered for each plate.
}
\section{Functions}{
\itemize{
\item \code{filterImages}: Filter images (rows) with a powerloglog slope 
value below the threshold ('num_IQR' * IQR below the 25th percentile) in 
ANY channel.

\item \code{plotFiltered}: Creates bar graph of the number of images filtered 
for each plate.
}}

