% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_all.R
\name{norm_all}
\alias{norm_all}
\title{Normalise data}
\usage{
norm_all(df, data_cols, count_col = "Count_Cells", compound_col, negCtrl,
  rescale_type = "robust", plate_col = "Metadata_Barcode")
}
\arguments{
\item{df}{Dataframe containing the CellProfiler data to normalise.}

\item{data_cols}{Numeric vector. Indicies of the data columns of df. All of 
these columns should be of numerice type. All remaining (non 'data_cols')
should NOT be of numeric type.}

\item{count_col}{Name of the column containing the raw cell counts, as 
string.}

\item{compound_col}{Name of the column in 'df' containing the compound 
details.}

\item{negCtrl}{Name of the negative control compound.}

\item{rescale_type}{Type of rescaling to use, as string. Will be passed to 
\code{reshape::rescaler}.}

\item{plate_col}{Name of the column containing the plate IDs.}
}
\value{
A dataframe with the following columns, (in order): 
\itemize{
    \item Annotation columns (all non-data columns).
    \item Cell count columns: raw cell count, cell count normalised to 
    negative control, z score normalised cell count (using 
    \code{reshape::rescaler}).
    \item All remaining data columns z score normalised (using 
    \code{reshape::rescaler}). All of these columns names will be the same 
    as the input dataframe.
}
}
\description{
Takes in CellProfiler morphology data and uses \code{reshape::rescaler} to 
normalise all data columns. Also calculates count value normalised to the
negative control, for each plate, and retains the raw count value.
}
