% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_negCtrl.R
\name{plot_negCtrl}
\alias{plot_negCtrl}
\title{Plot distribution of negative control wells}
\usage{
plot_negCtrl(df, feature, annot, neg_ctrl)
}
\arguments{
\item{df}{Dataframe of Cell Profiler data including well annotations.}

\item{feature}{Name of the feature column to plot as string.}

\item{annot}{Name of the column containing the annotation data as string.}

\item{neg_ctrl}{The term labelling the negative control wells in the 
annotation (e.g. "DMSO"), as string.}
}
\value{
Creates a plot in the current graphics device.
}
\description{
Create boxplot of the negative control wells for each plate for a given 
feature.
}
