% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plls.R
\name{plot_plls}
\alias{plot_plls}
\title{Plot powerloglog slope}
\usage{
plot_plls(df, channel, scale)
}
\arguments{
\item{df}{Dataframe containing a 'ImageQuality_PowerLogLogSlope' column}

\item{channel}{Name of the channel to be plotted, exactly as written in
Cell Profiler output, as string.}

\item{scale}{Single logical. If TRUE the scale of all plots will be uniform.
The x-axis limits will be set to the min and max of the powerloglog slope 
for all plates.}
}
\description{
Plot the density of the image quality metric, powerloglog slope for a 
specified channel, for each plate, in the current graphics device.
}
